guard against empty files for delete

+++ just_start_scraping.py
@@ -238,17 +238,19 @@ class JustStartSraping:
                 # date hasn't changes since the last row was written
                 if outpath.exists():
                     with outpath.open('r', newline='', encoding='utf8') as file:
-                        first_date = file.readlines()[-1].split(',')[0]
-                        current_date = race.date.strftime("%d/%m/%y")
-                        if first_date != current_date:
-                            # Must be the next day
-                            # Close file to avoid permission errors
-                            file.close()
-                            try:
-                                outpath.unlink()
-                            except:
-                                print('could not remove {}. It must be in use'.format(
-                                    str(outpath)))
+                        lines = file.readlines()
+                        if len(lines) > 0:
+                            first_date = lines[-1].split(',')[0]
+                            current_date = race.date.strftime("%d/%m/%y")
+                            if first_date != current_date:
+                                # Must be the next day
+                                # Close file to avoid permission errors
+                                file.close()
+                                try:
+                                    outpath.unlink()
+                                except:
+                                    print('could not remove {}. It must be in use'.format(
+                                        str(outpath)))
                 # now really write
                 with outpath.open('a', newline='', encoding='utf8') as file:
                     for r in runs:
